refactor(backend): refactor ContactListAPI to ports and adapters
