--- packages/backend/src/modules/marketing/ContactListAPI.ts
@@ -1,7 +1,0 @@
-export class ContactListAPI {
-  // will not implement communicating with a real api for the time being
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async addEmailToList(_email: string): Promise<void> {
-    return;
-  }
-}
