+++ packages/backend/src/modules/marketing/ports/ContactListAPI.ts
@@ -1,0 +1,3 @@
+export interface ContactListAPI {
+  addEmailToList(email: string): Promise<void>;
+}
